separated strucs related to reminder button to separate files
